Additional information is now extracted, and the API has been updated to make this available.
For example, S10 provides a table of information for each courier. This is now available via the courier_info property on TrackingNumber.

If a additional validator is provided, that is now correctly used, instead of the old behaviour that would fail if the additional information was missing. This matches the behaviour of the upstream ruby library, and makes the library more generous at parsing valid tracking numbers, that we don't have additional information for.